move service interfaces into /interfaces
