+++ src/MauiScreenTime/Services/DatabaseService.cs
@@ -1,3 +1,4 @@
+using MauiScreenTime.Services.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
