+++ src/MauiScreenTime/Services/Interfaces/ICO2Service.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MauiScreenTime.Services
+namespace MauiScreenTime.Services.Interfaces
 {
     public interface ICO2Service
     {
