+++ src/MauiScreenTime/Services/StartupService.cs
@@ -2,6 +2,7 @@
 using MauiScreenTime.Data;
 using MauiScreenTime.Data.Interfaces;
 using MauiScreenTime.Pages;
+using MauiScreenTime.Services.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
