+++ src/MauiScreenTime/Services/UsageStatsService.cs
@@ -1,4 +1,5 @@
 ﻿using MauiScreenTime.Data;
+using MauiScreenTime.Services.Interfaces;
 using Microsoft.Maui.Controls.PlatformConfiguration;
 using System;
 using System.Collections.Generic;
@@ -9,6 +10,7 @@ using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
 using static Microsoft.Maui.ApplicationModel.Permissions;
+
 
 #if ANDROID
 using Android.Content.PM;
